requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
databases==0.8.0
Django==4.2.6
django-phonenumber-field==7.2.0
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
phonenumberslite==8.13.22
Pillow==10.1.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
# mysqlclient==2.2.0
python-dotenv==1.0.0
daphne==4.0.0
channels==4.0.0
